backend/requirements.txt
# ── Core ──────────────────────────────────
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-multipart==0.0.*
python-dotenv==1.1.*

# ── Image handling ────────────────────────
Pillow==11.*

# ── AI / ML ───────────────────────────────
torch>=2.0
torchvision>=0.15
ultralytics==8.*
openai==1.*
google-genai>=1.0
numpy

# ── PDF generation ────────────────────────
fpdf2==2.*

# ── Utilities ─────────────────────────────
aiofiles==24.*
